askbot_requirements.txt
Coffin>=0.3,<=0.3.8
Jinja2==2.8
django-appconf
akismet<=0.2.0
beautifulsoup4<=4.4.1
django-avatar==2.2.1
django-celery>=3.0.11,<=3.1.17
django-compressor>=1.3,<=1.5
celery==3.1.18
django-countries==3.3
django-followit==0.2.1
django-keyedcache<=1.5.1
django-kombu==0.9.4
django-picklefield==0.3.0
django-recaptcha>=1.0.3,<=1.0.5
django-robots==1.1
django-threaded-multihost<=1.4-1
django-tinymce==1.5.3
django-transaction-signals==1.0.0
django>=1.8,<1.9
html5lib==0.9999999
jsonfield<=1.0.3
markdown2<=2.3.1
mock==1.0.1
oauth2<=1.9.0.post1
pyjwt<=1.4.0
pystache==0.3.1
python-cas==1.1.0
python-openid<=2.2.5
pytz<=2016.4
requirements-parser==0.1.0
regex
sanction==0.3.1
simplejson<=3.8.2
unidecode<=0.4.19
